sample_crud_task/requirements.txt
asgiref==3.6.0
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==1.7.1
pytz==2022.7
sqlparse==0.4.3
